Replace original press page with new one
Add About section

Unnecessary check

Use CMS "about"
