+++ src/app/pages/press/banner/banner.scss
