+++ src/app/pages/press/faq/faq.js
