+++ src/app/pages/press/featured-in/featured-in.scss
@@ -4,8 +4,14 @@
         gap: 2rem;
         grid-template-columns: repeat(auto-fit, minmax(16rem, 1fr));
 
+        > a {
+            align-items: center;
+            display: flex;
+            justify-content: center;
+            max-height: 16rem;
+        }
+
         img {
-            min-width: 16rem;
             max-width: 100%;
             max-height: 100%;
         }
