--- src/app/pages/press/inquiries/inquiries.scss
@@ -1,58 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-.press.page .inquiries {
-    display: grid;
-    grid-gap: 5rem;
-
-    > div {
-        display: grid;
-        grid-gap: 2rem;
-    }
-
-    .contact {
-        > :first-child {
-            font-weight: bold;
-        }
-    }
-
-    h2 {
-        @include set-font(h3);
-
-        margin-bottom: 0;
-    }
-
-    button,
-    .btn {
-        @extend %button;
-
-        display: grid;
-        grid-gap: 1rem;
-        height: auto;
-        justify-items: center;
-        justify-self: start;
-        padding-bottom: 0.5rem;
-
-        &.primary {
-            @extend %primary;
-
-            height: auto;
-        }
-    }
-
-    .icon-row {
-        color: os-color(green);
-        display: grid;
-        grid-gap: 1rem;
-        grid-template-columns: repeat(4, 3.1rem);
-        height: 3.6rem;
-
-        > a {
-            color: inherit;
-
-            > * {
-                height: 100%;
-                width: 100%;
-            }
-        }
-    }
-}
