+++ src/app/pages/press/mission/leadership.scss
@@ -11,7 +11,7 @@
     }
 
     .bio {
-        max-width: 40rem;
+        max-width: 80rem;
 
         .card {
             overflow: hidden;
