--- src/app/pages/press/mobile-selector/mobile-selector.scss
@@ -1,51 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-.mobile-selector {
-    padding: 1rem 2rem 0;
-
-    .selector-button {
-        @extend %button;
-
-        background-color: ui-color(white);
-        justify-content: space-between;
-        width: 100%;
-    }
-
-    .fixed-overlay {
-        align-items: center;
-        background-color: rgba(ui-color(black), 0.5);
-        display: grid;
-        height: 100vh;
-        justify-items: center;
-        left: 0;
-        position: fixed;
-        top: 0;
-        width: 100vw;
-        z-index: 1;
-
-        [role="menu"] {
-            @include body-font(2.4rem);
-
-            background-color: ui-color(white);
-            border-radius: 0.25rem;
-            width: calc(100% - #{2 * $normal-margin});
-        }
-
-        [role="menuitem"] {
-            padding: 1rem $normal-margin;
-
-            &.selected {
-                background-color: ui-color(page-bg);
-                border-left: 0.3rem solid os-color(green);
-            }
-
-            &:first-child {
-                border-top-right-radius: inherit;
-            }
-
-            &:last-child {
-                border-bottom-right-radius: inherit;
-            }
-        }
-    }
-}
