--- src/app/pages/press/press-excerpt/press-excerpt.scss
@@ -1,42 +1,0 @@
-@import 'pattern-library/core/pattern-library/headers';
-
-.press-excerpt {
-    display: grid;
-
-    &.with-icon {
-        grid-column-gap: 2rem;
-        grid-template-columns: 5rem 1fr;
-    }
-
-    > img {
-        grid-row-start: span 2;
-        max-width: 100%;
-    }
-
-    .byline {
-        @include set-font(helper-label);
-    }
-
-    .source {
-        font-style: italic;
-    }
-
-    .headline {
-        font-weight: bold;
-
-        a {
-            display: block;
-            text-decoration: none;
-
-            &:not(:hover) {
-                color: inherit;
-            }
-
-            .fa-external-link-alt {
-                font-size: 75%;
-                margin-bottom: 0.1rem;
-                margin-left: 0.1rem;
-            }
-        }
-    }
-}
