+++ src/app/pages/press/press.scss
@@ -1,128 +1,70 @@
 @import 'pattern-library/core/pattern-library/headers';
 
-.press.page {
+.page.press {
     background-color: ui-color(white);
-    padding-bottom: 2rem;
+    display: grid;
+    gap: 4rem 3rem;
+    padding-bottom: 6rem;
 
-    .hero {
-        background-image: linear-gradient(to left, lighten(os-color(green), 18%), darken(os-color(green), 3%));
-        padding: 14rem 0 5rem;
+    section {
+        max-width: calc(100vw - 3rem);
 
-        @include width-up-to($phone-max) {
-            padding-top: 6rem;
+        @include width-up-to($tablet-max) {
+            margin: 0 $normal-margin;
         }
 
-        h1 {
-            @extend %content;
-
-            color: text-color(white);
-            letter-spacing: normal;
-            text-align: left;
-
-            @include width-up-to($phone-max) {
-                @include title-font(3.6rem);
-            }
+        .boxed {
+            align-items: start;
         }
     }
 
-    h2 {
-        @include width-up-to($phone-max) {
-            @include title-font(2.4rem);
+    @include wider-than($tablet-max) {
+        grid-template-columns: calc(50vw - 60rem) 2fr 1fr 2fr calc(50vw - 60rem);
 
-            margin-bottom: 3rem;
+        #in-the-press {
+            grid-column: 1 / 6;
+        }
+
+        #testimonials,
+        #faq {
+            grid-column: 2 / 5;
+        }
+
+        #featured-in,
+        #facts-infographic,
+        #about-openstax {
+            grid-column: 2 / 4;
+        }
+
+        #press-releases {
+            grid-column: 2 / 3;
+        }
+
+        #news-mentions {
+            grid-column: 3 / 5;
+        }
+
+        #mission {
+            grid-column: 4 / 5;
+            grid-row: span 3;
         }
     }
 
-    .mobile-selector {
-        @include wider-than($phone-max) {
-            display: none;
-        }
-    }
-
-    .main-body {
+    .content-block {
         @extend %content;
+        @include scale-feature(gap, $normal-margin, 2.5rem);
+        padding: 0;
 
-        display: grid;
-        grid-column-gap: calc(100% / 6);
-        grid-template-columns: 3fr 2fr;
-        padding-top: 4rem;
+        display: flex;
+        flex-direction: column;
+        max-width: 100%;
 
-        @include width-up-to($phone-max) {
-            grid-template-columns: 1fr;
-            margin-bottom: 2rem;
-
-            .hidden-on-mobile:not(.active) {
-                display: none;
-            }
-        }
-
-        @include wider-than($phone-max) {
-            .mobile-only {
-                display: none;
-            }
-        }
-    }
-
-    .main-content,
-    .sidebar {
-        display: grid;
-        grid-row-gap: 6rem;
-
-        @include width-up-to($phone-max) {
-            display: block;
-        }
-
-        h2 {
-            line-height: normal;
+        > * {
             margin: 0;
         }
     }
 
-    .main-content {
-        grid-template-rows: auto auto 1fr;
-    }
-
-    .sidebar {
-        grid-template-rows: auto auto auto 1fr;
-    }
-
-    .our-mission,
-    .book-experts {
-        display: grid;
-        grid-row-gap: 2rem;
-
-        h2 {
-            @include set-font(h3);
-        }
-    }
-
-    [data-region="bookings"] {
-        display: grid;
-        grid-row-gap: 4rem;
-    }
-
-    .press-releases,
-    .news-mentions {
-        display: grid;
-        grid-row-gap: 2rem;
-        grid-template-rows: auto auto;
-    }
-
-    .hero + article,
-    .other.press-releases,
-    .text-content > h1 {
-        width: 100%;
-    }
-
-    .article {
-        figure {
-            @include width-up-to($phone-max) {
-                margin-inline: 2rem;
-            }
-
-            img {
-                max-width: 100%;
-            }
-        }
+    h2 {
+        margin: 0;
     }
 }
